requirements.txt
amqp==5.2.0
asgiref==3.7.2
asttokens==2.4.1
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.3.3
cron-descriptor==1.4.0
decorator==5.1.1
Django==4.2.8
django-celery-beat==2.5.0
django-environ==0.11.2
django-filter==23.5
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
executing==2.0.1
idna==3.6
inflection==0.5.1
ipython==8.18.1
jedi==0.19.1
kombu==5.3.4
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
stripe==7.9.0
traitlets==5.14.0
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
